+++ smol-bytes/src/lib.rs
@@ -50,6 +50,10 @@ mod bytes;
 #[cfg(any(feature = "std", feature = "alloc"))]
 mod bytes_mut;
 
+#[cfg(feature = "async-graphql")]
+#[cfg_attr(docsrs, doc(cfg(feature = "async-graphql")))]
+mod graphql;
+
 #[cfg(feature = "pyo3")]
 #[cfg_attr(docsrs, doc(cfg(feature = "pyo3")))]
 mod python;